Refactor, prepping for playoffs row
