+++ src/components/MultiDisplay/EventRow/sharedStyles.module.scss
@@ -53,43 +53,3 @@ tbody > tr {
 .w100 {
     width: 100%;
 }
-
-.messageContainer {
-    transition: all ease .5s;
-    position: relative;
-    z-index: 3;
-    height: 0;
-    width: 0;
-    left: -101vw;
-}
-
-.messageMover {
-    width: 100vw;
-    height: 22vh;
-    background-color: black;
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-
-    &>span {
-        width: 80%;
-    }
-}
-
-.faderContainer {
-    width: 30vw;
-    height: 22vh;
-
-    &>* {
-        position: absolute;
-        transition: all ease .5s;
-        text-align: center;
-        width: 30vw;
-        top: 0;
-        font-weight: bold;
-        align-items: center;
-        justify-content: center;
-    };
-
-}
